Create python-app.yml
in the "on " section of the yml, we specified that the workflow should trigger on push events on the main branch with types closed. This means that it will trigger when a pull request that targets the main branch is closed.

In the jobs.build section, we added an if condition to only run the job if the pull request has been merged (github.event.pull_request.merged == true). This ensures that the workflow will only run if the pull request was successfully merged.